+++ lib/cucumber/core/report/summary.rb
@@ -4,22 +4,45 @@ module Cucumber
       class Summary
         attr_reader :test_cases, :test_steps
 
-        def initialize
+        def initialize(event_bus)
           @test_cases = Test::Result::Summary.new
           @test_steps = Test::Result::Summary.new
+          subscribe_to(event_bus)
         end
 
-        def after_test_case(test_case, result)
-          result.describe_to test_cases
+        private
+
+        def subscribe_to(event_bus)
+          event_bus.on(:test_case_finished) do |event|
+            event.result.describe_to test_cases
+          end
+          event_bus.on(:test_step_finished) do |event|
+            event.result.describe_to test_steps if is_step?(event.test_step)
+          end
+          self
         end
 
-        def after_test_step(test_step, result)
-          result.describe_to test_steps
+        def is_step?(test_step)
+          StepQueryVisitor.new(test_step).is_step?
+        end
+      end
+
+      class StepQueryVisitor
+        def initialize(test_step)
+          @step = false
+          test_step.source.last.describe_to(self)
         end
 
+        def is_step?
+          @step
+        end
+
+        def step(*)
+          @step = true
+        end
+        
         def method_missing(*)
         end
-
       end
     end
   end