+++ spec/cucumber/core/test/runner_spec.rb
@@ -8,8 +8,8 @@ module Cucumber::Core::Test
 
     let(:test_case) { Case.new(test_steps, source) }
     let(:source)    { [double('ast node', location: double)] }
-    let(:runner)    { Runner.new(report) }
-    let(:report)    { double.as_null_object }
+    let(:runner)    { Runner.new(event_bus) }
+    let(:event_bus)    { double.as_null_object }
     let(:passing)   { Step.new([source]).with_action {} }
     let(:failing)   { Step.new([source]).with_action { raise exception } }
     let(:pending)   { Step.new([source]).with_action { raise Result::Pending.new("TODO") } }
@@ -18,7 +18,7 @@ module Cucumber::Core::Test
     let(:exception) { StandardError.new('test error') }
 
     before do
-      allow(report).to receive(:before_test_case)
+      allow(event_bus).to receive(:test_case_starting)
       allow(source).to receive(:location)
     end
 
@@ -31,7 +31,7 @@ module Cucumber::Core::Test
         let(:test_steps) { [passing] }
 
         it "records the nanoseconds duration of the execution on the result" do
-          expect( report ).to receive(:after_test_case) do |reported_test_case, result|
+          expect(event_bus).to receive(:test_case_finished) do |reported_test_case, result|
             expect( result.duration ).to be_duration 1
           end
           test_case.describe_to runner
@@ -42,8 +42,8 @@ module Cucumber::Core::Test
         let(:test_steps) { [failing] }
 
         it "records the duration" do
-          expect( report ).to receive(:after_test_case) do |reported_test_case, result|
-            expect( result.duration ).to be_duration 1
+          expect(event_bus).to receive(:test_case_finished) do |reported_test_case, result|
+            expect(result.duration).to be_duration 1
           end
           test_case.describe_to runner
         end
@@ -53,8 +53,8 @@ module Cucumber::Core::Test
     context "reporting the exception that failed a test case" do
       let(:test_steps) { [failing] }
       it "sets the exception on the result" do
-        allow(report).to receive(:before_test_case)
-        expect( report ).to receive(:after_test_case) do |reported_test_case, result|
+        allow(event_bus).to receive(:before_test_case)
+        expect(event_bus).to receive(:test_case_finished) do |reported_test_case, result|
           expect( result.exception ).to eq exception
         end
         test_case.describe_to runner
@@ -65,13 +65,13 @@ module Cucumber::Core::Test
       context "without steps" do
         let(:test_steps) { [] }
 
-        it "calls the report before running the case" do
-          expect( report ).to receive(:before_test_case).with(test_case)
+        it "emits a test_case_starting event before running the test case" do
+          expect(event_bus).to receive(:test_case_starting).with(test_case)
           test_case.describe_to runner
         end
 
-        it "calls the report after running the case" do
-          expect( report ).to receive(:after_test_case) do |reported_test_case, result|
+        it "emits the test_case_finished event after running the the test case" do
+          expect(event_bus).to receive(:test_case_finished) do |reported_test_case, result|
             expect( reported_test_case ).to eq test_case
             expect( result ).to be_unknown
           end
@@ -83,8 +83,8 @@ module Cucumber::Core::Test
         context 'that all pass' do
           let(:test_steps) { [ passing, passing ]  }
 
-          it 'reports a passing test case' do
-            expect( report ).to receive(:after_test_case) do |test_case, result|
+          it 'emits the test_case_finished event with a passing result' do
+            expect(event_bus).to receive(:test_case_finished) do |test_case, result|
               expect( result ).to be_passed
             end
             test_case.describe_to runner
@@ -94,8 +94,8 @@ module Cucumber::Core::Test
         context 'an undefined step' do
           let(:test_steps) { [ undefined ]  }
 
-          it 'reports an undefined test case' do
-            expect( report ).to receive(:after_test_case) do |test_case, result|
+          it 'emits the test_case_finished event with an undefined result' do
+            expect(event_bus).to receive(:test_case_finished) do |test_case, result|
               expect( result ).to be_undefined
             end
             allow( undefined.source.last ).to receive(:name)
@@ -103,7 +103,7 @@ module Cucumber::Core::Test
           end
 
           it 'sets the message on the result' do
-            expect( report ).to receive(:after_test_case) do |test_case, result|
+            expect(event_bus).to receive(:test_case_finished) do |test_case, result|
               expect( result.message ).to eq("Undefined step: \"step name\"")
             end
             expect( undefined.source.last ).to receive(:name).and_return("step name")
@@ -111,7 +111,7 @@ module Cucumber::Core::Test
           end
 
           it 'appends the backtrace of the result' do
-            expect( report ).to receive(:after_test_case) do |test_case, result|
+            expect(event_bus).to receive(:test_case_finished) do |test_case, result|
               expect( result.backtrace ).to eq(["step line"])
             end
             expect( undefined.source.last ).to receive(:backtrace_line).and_return("step line")
@@ -123,15 +123,15 @@ module Cucumber::Core::Test
         context 'a pending step' do
           let(:test_steps) { [ pending ] }
 
-          it 'reports a pending test case' do
-            expect( report ).to receive(:after_test_case) do |test_case, result|
+          it 'emits the test_case_finished event with a pending result' do
+            expect(event_bus).to receive(:test_case_finished) do |test_case, result|
               expect( result ).to be_pending
             end
             test_case.describe_to runner
           end
 
           it 'appends the backtrace of the result' do
-            expect( report ).to receive(:after_test_case) do |test_case, result|
+            expect(event_bus).to receive(:test_case_finished) do |test_case, result|
               expect( result.backtrace.last ).to eq("step line")
             end
             expect( pending.source.last ).to receive(:backtrace_line).and_return("step line")
@@ -142,15 +142,15 @@ module Cucumber::Core::Test
         context "a skipping step" do
           let(:test_steps) { [skipping] }
 
-          it "reports a skipped test case" do
-            expect( report ).to receive(:after_test_case) do |test_case, result|
+          it "emits the test_case_finished event with a skipped result" do
+            expect(event_bus).to receive(:test_case_finished) do |test_case, result|
               expect( result ).to be_skipped
             end
             test_case.describe_to runner
           end
 
           it 'appends the backtrace of the result' do
-            expect( report ).to receive(:after_test_case) do |test_case, result|
+            expect(event_bus).to receive(:test_case_finished) do |test_case, result|
               expect( result.backtrace.last ).to eq("step line")
             end
             expect( skipping.source.last ).to receive(:backtrace_line).and_return("step line")
@@ -161,15 +161,15 @@ module Cucumber::Core::Test
         context 'that fail' do
           let(:test_steps) { [ failing ] }
 
-          it 'reports a failing test case' do
-            expect( report ).to receive(:after_test_case) do |test_case, result|
+          it 'emits the test_case_finished event with a failing result' do
+            expect(event_bus).to receive(:test_case_finished) do |test_case, result|
               expect( result ).to be_failed
             end
             test_case.describe_to runner
           end
 
           it 'appends the backtrace of the exception of the result' do
-            expect( report ).to receive(:after_test_case) do |test_case, result|
+            expect(event_bus).to receive(:test_case_finished) do |test_case, result|
               expect( result.exception.backtrace.last ).to eq("step line")
             end
             expect( failing.source.last ).to receive(:backtrace_line).and_return("step line")
@@ -180,30 +180,22 @@ module Cucumber::Core::Test
         context 'where the first step fails' do
           let(:test_steps) { [ failing, passing ] }
 
-          it 'executes the after hook at the end regardless of the failure' do
-            expect( report ).to receive(:after_test_case) do |test_case, result|
-              expect( result ).to be_failed
-              expect( result.exception ).to eq exception
-            end
-            test_case.describe_to runner
-          end
-
-          it 'reports the first step as failed' do
-            expect( report ).to receive(:after_test_step).with(failing, anything) do |test_step, result|
+          it 'emits the test_step_finished event with a failed result' do
+            expect(event_bus).to receive(:test_step_finished).with(failing, anything) do |test_step, result|
               expect( result ).to be_failed
             end
             test_case.describe_to runner
           end
 
-          it 'reports the second step as skipped' do
-            expect( report ).to receive(:after_test_step).with(passing, anything) do |test_step, result|
+          it 'emits a test_step_finished event with a skipped result' do
+            expect(event_bus).to receive(:test_step_finished).with(passing, anything) do |test_step, result|
               expect( result ).to be_skipped
             end
             test_case.describe_to runner
           end
 
-          it 'reports the test case as failed' do
-            expect( report ).to receive(:after_test_case) do |test_case, result|
+          it 'emits a test_case_finished event with a failed result' do
+            expect(event_bus).to receive(:test_case_finished) do |test_case, result|
               expect( result ).to be_failed
               expect( result.exception ).to eq exception
             end
@@ -227,7 +219,7 @@ module Cucumber::Core::Test
         let(:test_cases)      { [first_test_case, last_test_case] }
 
         it 'reports the results correctly for the following test case' do
-          expect( report ).to receive(:after_test_case).with(last_test_case, anything) do |reported_test_case, result|
+          expect(event_bus).to receive(:test_case_finished).with(last_test_case, anything) do |reported_test_case, result|
             expect( result ).to be_passed
           end
 
@@ -256,7 +248,7 @@ module Cucumber::Core::Test
         around_hook = AroundHook.new { |block| block.call }
         passing_step = Step.new([source]).with_action {}
         test_case = Case.new([passing_step], source, [around_hook])
-        expect(report).to receive(:after_test_case).with(test_case, anything) do |reported_test_case, result|
+        expect(event_bus).to receive(:test_case_finished).with(test_case, anything) do |reported_test_case, result|
           expect(result).to be_passed
         end
         test_case.describe_to runner
@@ -266,7 +258,7 @@ module Cucumber::Core::Test
         around_hook = AroundHook.new { |block| raise exception }
         passing_step = Step.new([source]).with_action {}
         test_case = Case.new([passing_step], source, [around_hook])
-        expect(report).to receive(:after_test_case).with(test_case, anything) do |reported_test_case, result|
+        expect(event_bus).to receive(:test_case_finished).with(test_case, anything) do |reported_test_case, result|
           expect(result).to be_failed
           expect(result.exception).to eq exception
         end
@@ -277,7 +269,7 @@ module Cucumber::Core::Test
         around_hook = AroundHook.new { |block| block.call; raise exception }
         passing_step = Step.new([source]).with_action {}
         test_case = Case.new([passing_step], source, [around_hook])
-        expect(report).to receive(:after_test_case).with(test_case, anything) do |reported_test_case, result|
+        expect(event_bus).to receive(:test_case_finished).with(test_case, anything) do |reported_test_case, result|
           expect(result).to be_failed
           expect(result.exception).to eq exception
         end
@@ -288,7 +280,7 @@ module Cucumber::Core::Test
         around_hook = AroundHook.new { |block| block.call }
         failing_step = Step.new([source]).with_action { raise exception }
         test_case = Case.new([failing_step], source, [around_hook])
-        expect(report).to receive(:after_test_case).with(test_case, anything) do |reported_test_case, result|
+        expect(event_bus).to receive(:test_case_finished).with(test_case, anything) do |reported_test_case, result|
           expect(result).to be_failed
           expect(result.exception).to eq exception
         end
@@ -300,11 +292,11 @@ module Cucumber::Core::Test
         passing_step = Step.new([source]).with_action {}
         test_case = Case.new([], source, [around_hook])
         allow(runner).to receive(:running_test_step).and_return(passing_step)
-        expect(report).to receive(:after_test_step).with(passing_step, anything) do |reported_test_case, result|
+        expect(event_bus).to receive(:test_step_finished).with(passing_step, anything) do |reported_test_case, result|
           expect(result).to be_failed
           expect(result.exception).to eq exception
         end
-        expect(report).to receive(:after_test_case).with(test_case, anything) do |reported_test_case, result|
+        expect(event_bus).to receive(:test_case_finished).with(test_case, anything) do |reported_test_case, result|
           expect(result).to be_failed
           expect(result.exception).to eq exception
         end